+++ e2e-tests/ui-components/playwright.config.js
@@ -1,4 +1,6 @@
 import { defineConfig, devices } from "@playwright/test";
+import dotenv from "dotenv";
+dotenv.config({ path: "../../.env" });
 
 /**
  * Read environment variables from file.