+++ packages/react-native/src/components/Card/Card.tsx
@@ -107,19 +107,6 @@ function Card({ initialValue, config, children, onChange, style }: CardProps) {
     },
   });
 
-  useEffect(() => {
-    const getCardData = async () => {
-      const cardData = await changePayload(
-        encrypt,
-        form,
-        Array.from(registeredFields)
-      );
-    };
-    if (isComplete(form, Array.from(registeredFields))) {
-      getCardData();
-    }
-  }, [form, registeredFields]);
-
   return (
     <CardContext.Provider
       value={{