+++ packages/react/lib/ui/Card.tsx
@@ -4,6 +4,7 @@ import { useEffect, useLayoutEffect, useMemo, useRef } from "react";
 import { useEvervault } from "../useEvervault";
 import type {
   CardField,
+  CardOptions,
   CardPayload,
   CardTranslations,
   SwipedCard,
@@ -20,6 +21,8 @@ export interface CardProps {
   onSwipe?: (data: SwipedCard) => void;
   onChange?: (data: CardPayload) => void;
   onComplete?: (data: CardPayload) => void;
+  autoComplete?: CardOptions["autoComplete"];
+  autoProgress?: boolean;
 }
 
 type CardClass = ReturnType<Evervault["ui"]["card"]>;
@@ -34,6 +37,8 @@ export function Card({
   onError,
   onChange,
   onComplete,
+  autoComplete,
+  autoProgress,
 }: CardProps) {
   const ev = useEvervault();
   const initialized = useRef(false);
@@ -76,8 +81,10 @@ export function Card({
       fields,
       autoFocus,
       translations,
+      autoComplete,
+      autoProgress,
     }),
-    [theme, translations, fields, autoFocus]
+    [theme, translations, fields, autoFocus, autoComplete, autoProgress]
   );
 
   useLayoutEffect(() => {