+++ packages/ui-components/src/Card/CardCVC.tsx
@@ -1,4 +1,4 @@
-import {validateNumber} from "@evervault/card-validator";
+import { validateNumber } from "@evervault/card-validator";
 import {
   FocusEvent,
   forwardRef,
@@ -17,11 +17,21 @@ interface CVCProps {
   value: string;
   readOnly?: boolean;
   cardNumber: string;
+  autoComplete?: boolean;
 }
 
 export const CardCVC = forwardRef<HTMLInputElement, CVCProps>(
   (
-    { cardNumber, onChange, onBlur, disabled, placeholder, value, readOnly },
+    {
+      cardNumber,
+      onChange,
+      onBlur,
+      disabled,
+      placeholder,
+      value,
+      readOnly,
+      autoComplete,
+    },
     forwardedRef
   ) => {
     const innerRef = useRef<HTMLInputElement>(null);
@@ -29,6 +39,7 @@ export const CardCVC = forwardRef<HTMLInputElement, CVCProps>(
     useImperativeHandle(forwardedRef, () => innerRef.current!);
 
     const mask = useMemo(() => {
+      if (!cardNumber) return "0000";
       const type = validateNumber(cardNumber).brand;
       if (type === "american-express") return "0000";
       return "000";
@@ -50,7 +61,7 @@ export const CardCVC = forwardRef<HTMLInputElement, CVCProps>(
         onBlur={onBlur}
         placeholder={placeholder}
         pattern="[0-9]*"
-        autoComplete="billing cc-cvc"
+        autoComplete={autoComplete ? "billing cc-cvc" : "off"}
         readOnly={readOnly}
       />
     );