+++ packages/ui-components/src/ThreeDSecure/BrandLogo/Amex.tsx
@@ -1,0 +1,51 @@
+export function Amex() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="60"
+      height="60"
+      fill="none"
+      viewBox="0 0 60 60"
+    >
+      <g clipPath="url(#clip0_32_36)">
+        <path
+          fill="color(display-p3 0.0000 0.4353 0.8118)"
+          d="M0 60h60V0H0v60z"
+        ></path>
+        <path
+          fill="#fff"
+          d="M21.735 50.506l23.725.05 3.726-4.156 3.752 4.14H60v-6.064l-4.423-4.77L60 34.98v-5.939l-38.265-.034v21.499z"
+        ></path>
+        <path
+          fill="color(display-p3 0.0000 0.4353 0.8118)"
+          fillRule="evenodd"
+          d="M24.291 48.149V31.373h14.225v3.864h-9.62v2.624h9.39v3.8h-9.39v2.577h9.62v3.91l-14.225.001z"
+          clipRule="evenodd"
+        ></path>
+        <path
+          fill="color(display-p3 0.0000 0.4353 0.8118)"
+          d="M54.17 31.373l-4.824 5.317-4.81-5.317h-6.092l7.871 8.378-7.87 8.398h5.926l4.845-5.37 4.895 5.37H60v-.238l-7.703-8.16L60 31.505v-.132h-5.83z"
+        ></path>
+        <path
+          fill="#fff"
+          d="M51.305 9.925l-1.97 5.511-1.964-5.511H35.984v7.356l-3.24-7.356H23.52l-9.473 21.499H60v-21.5h-8.695z"
+        ></path>
+        <path
+          fill="color(display-p3 0.0000 0.4353 0.8118)"
+          fillRule="evenodd"
+          d="M25.906 21.904l2.232-5.336 2.231 5.336h-4.463zm-.64-9.644l-7.443 16.762h5.105l1.404-3.356h7.61l1.403 3.356h5.232L31.165 12.26h-5.899z"
+          clipRule="evenodd"
+        ></path>
+        <path
+          fill="color(display-p3 0.0000 0.4353 0.8118)"
+          d="M38.507 29.02h4.514V17.505l4.282 11.513h4.057l4.195-11.474v11.513L60 29.02V12.258h-6.907l-3.717 10.33-3.692-10.306-7.177-.024v16.761z"
+        ></path>
+      </g>
+      <defs>
+        <clipPath id="clip0_32_36">
+          <path fill="#fff" d="M0 0H60V60H0z"></path>
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}