+++ packages/ui-components/src/ThreeDSecure/BrandLogo/JCB.tsx
@@ -1,0 +1,157 @@
+export function JCB() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="78"
+      height="60"
+      fill="none"
+      viewBox="0 0 78 60"
+    >
+      <g clipPath="url(#clip0_32_148)">
+        <path
+          fill="#fff"
+          d="M77.491 47.75c0 6.61-5.382 11.99-11.99 11.99H.258V12.25C.26 5.64 5.641.26 12.25.26h65.241v47.49z"
+        ></path>
+        <path
+          fill="url(#paint0_linear_32_148)"
+          d="M56.248 35.57h4.956c.142 0 .472-.047.614-.047.944-.189 1.747-1.038 1.747-2.219 0-1.133-.803-1.982-1.747-2.218-.142-.047-.425-.047-.614-.047h-4.956v4.531z"
+        ></path>
+        <path
+          fill="url(#paint1_linear_32_148)"
+          d="M60.638 4.272c-4.72 0-8.592 3.824-8.592 8.592v8.922H64.18c.283 0 .613 0 .85.047 2.737.142 4.767 1.558 4.767 4.013 0 1.935-1.369 3.587-3.918 3.918v.094c2.785.19 4.91 1.747 4.91 4.155 0 2.596-2.36 4.296-5.476 4.296H51.999v17.466h12.604c4.721 0 8.592-3.823 8.592-8.591V4.272H60.638z"
+        ></path>
+        <path
+          fill="url(#paint2_linear_32_148)"
+          d="M62.951 26.412c0-1.133-.803-1.888-1.747-2.03-.094 0-.33-.047-.472-.047h-4.484v4.154h4.484c.142 0 .425 0 .472-.047.944-.142 1.747-.897 1.747-2.03z"
+        ></path>
+        <path
+          fill="url(#paint3_linear_32_148)"
+          d="M13.147 4.272c-4.721 0-8.592 3.824-8.592 8.592V34.06c2.408 1.18 4.91 1.935 7.412 1.935 2.974 0 4.579-1.793 4.579-4.248V21.739h7.364v9.96c0 3.872-2.407 7.034-10.574 7.034-4.957 0-8.828-1.085-8.828-1.085v18.08h12.604c4.721 0 8.592-3.824 8.592-8.592V4.272H13.147z"
+        ></path>
+        <path
+          fill="url(#paint4_linear_32_148)"
+          d="M36.893 4.272c-4.721 0-8.592 3.824-8.592 8.592v11.235c2.171-1.841 5.948-3.021 12.038-2.738 3.257.142 6.75 1.038 6.75 1.038v3.636c-1.746-.897-3.823-1.7-6.514-1.889-4.627-.33-7.412 1.936-7.412 5.901 0 4.013 2.785 6.279 7.412 5.901 2.69-.189 4.768-1.038 6.514-1.888v3.635s-3.446.897-6.75 1.038c-6.09.284-9.867-.897-12.038-2.738v19.828h12.604c4.721 0 8.592-3.824 8.592-8.592V4.27H36.893z"
+        ></path>
+      </g>
+      <defs>
+        <linearGradient
+          id="paint0_linear_32_148"
+          x1="52.038"
+          x2="73.252"
+          y1="33.31"
+          y2="33.31"
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="color(display-p3 0.0000 0.4745 0.2510)"></stop>
+          <stop
+            offset="0.229"
+            stopColor="color(display-p3 0.0000 0.5294 0.2471)"
+          ></stop>
+          <stop
+            offset="0.743"
+            stopColor="color(display-p3 0.2510 0.6549 0.2157)"
+          ></stop>
+          <stop
+            offset="1"
+            stopColor="color(display-p3 0.3608 0.7098 0.1922)"
+          ></stop>
+        </linearGradient>
+        <linearGradient
+          id="paint1_linear_32_148"
+          x1="52.038"
+          x2="73.253"
+          y1="30.003"
+          y2="30.003"
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="color(display-p3 0.0000 0.4745 0.2510)"></stop>
+          <stop
+            offset="0.229"
+            stopColor="color(display-p3 0.0000 0.5294 0.2471)"
+          ></stop>
+          <stop
+            offset="0.743"
+            stopColor="color(display-p3 0.2510 0.6549 0.2157)"
+          ></stop>
+          <stop
+            offset="1"
+            stopColor="color(display-p3 0.3608 0.7098 0.1922)"
+          ></stop>
+        </linearGradient>
+        <linearGradient
+          id="paint2_linear_32_148"
+          x1="52.038"
+          x2="73.252"
+          y1="26.406"
+          y2="26.406"
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="color(display-p3 0.0000 0.4745 0.2510)"></stop>
+          <stop
+            offset="0.229"
+            stopColor="color(display-p3 0.0000 0.5294 0.2471)"
+          ></stop>
+          <stop
+            offset="0.743"
+            stopColor="color(display-p3 0.2510 0.6549 0.2157)"
+          ></stop>
+          <stop
+            offset="1"
+            stopColor="color(display-p3 0.3608 0.7098 0.1922)"
+          ></stop>
+        </linearGradient>
+        <linearGradient
+          id="paint3_linear_32_148"
+          x1="4.545"
+          x2="26.087"
+          y1="30.003"
+          y2="30.003"
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="color(display-p3 0.1216 0.1569 0.4353)"></stop>
+          <stop
+            offset="0.475"
+            stopColor="color(display-p3 0.0000 0.3059 0.5804)"
+          ></stop>
+          <stop
+            offset="0.826"
+            stopColor="color(display-p3 0.0000 0.4000 0.6941)"
+          ></stop>
+          <stop
+            offset="1"
+            stopColor="color(display-p3 0.0000 0.4353 0.7373)"
+          ></stop>
+        </linearGradient>
+        <linearGradient
+          id="paint4_linear_32_148"
+          x1="28.179"
+          x2="49.101"
+          y1="30.003"
+          y2="30.003"
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="color(display-p3 0.4235 0.1725 0.1843)"></stop>
+          <stop
+            offset="0.173"
+            stopColor="color(display-p3 0.5333 0.1529 0.1882)"
+          ></stop>
+          <stop
+            offset="0.573"
+            stopColor="color(display-p3 0.7451 0.0941 0.2000)"
+          ></stop>
+          <stop
+            offset="0.859"
+            stopColor="color(display-p3 0.8627 0.0157 0.2118)"
+          ></stop>
+          <stop
+            offset="1"
+            stopColor="color(display-p3 0.9020 0.0000 0.2235)"
+          ></stop>
+        </linearGradient>
+        <clipPath id="clip0_32_148">
+          <path fill="#fff" d="M0 0H77.75V60H0z"></path>
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}