+++ e2e-tests/browser/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @evervault/browser-e2e-tests
 
+## 1.0.3
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@1.0.3
+- @evervault/e2e-decrypt-backend@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@1.0.2
+- @evervault/e2e-decrypt-backend@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@1.0.1
+- @evervault/e2e-decrypt-backend@1.0.1
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null
+- @evervault/e2e-decrypt-backend@null
+
 ## null
 
 ### Patch Changes