+++ examples/collect-card-details/CHANGELOG.md
@@ -1,5 +1,34 @@
 # example-ui-components
 
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies [3156568]
+  - @evervault/browser@2.28.0
+
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies [1df9ba2]
+- Updated dependencies [e2f0721]
+  - @evervault/browser@2.27.0
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies [05eda8e]
+  - @evervault/browser@2.26.0
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [408c73e]
+  - @evervault/browser@2.25.2
+
 ## 0.0.14
 
 ### Patch Changes