+++ packages/react-native-example/CHANGELOG.md
@@ -1,5 +1,40 @@
 # react-native-example
 
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [e7d1bde]
+  - @evervault/evervault-react-native@0.5.3
+
+## 1.0.8
+
+### Patch Changes
+
+- 30c7f68: Fix cc-name completion on android
+- Updated dependencies [30c7f68]
+  - @evervault/evervault-react-native@0.5.2
+
+## 1.0.7
+
+### Patch Changes
+
+- @evervault/evervault-react-native@0.5.1
+
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [2070874]
+  - @evervault/evervault-react-native@0.5.0
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [4fda229]
+  - @evervault/evervault-react-native@0.4.6
+
 ## 1.0.4
 
 ### Patch Changes